Deprecate User.isEmailVerified
Determining if an email is verified from the Okta Credentials object is error-prone. On our oktapreview config the
credentials object no longer contains the emails for a user, which means we can't use the verified status it contains.
Because isEmailVerified was a concept used by The Key, and The Key is almost phased out, I'm opting to deprecate
isEmailVerified and default it to true for any place it's still used within The Key